feat: add fuzz targets, extensive tests, and CU benchmarks for liquidity math
- Add lib.rs benchmark entrypoints and argument structs for measuring compute units.
- Expose calculate_amount_0_delta and calculate_amount_1_delta as public for testing.
- Add comprehensive fuzz targets and dictionaries:
  - fuzz_calculate_amounts, fuzz_calculate_liquidity, fuzz_position_value
  - dictionaries for calculate_liquidity and position_value
  - update fuzz Cargo.toml and adjust Cargo.lock dependency ordering
- Introduce large suite of tests and regressions:
  - proptest regressions and expanded property tests (liquidity_property_tests)
  - comprehensive unit tests for liquidity math (huge unit test file)
  - new integration and economic-soundness test suites
  - wire unit_tests mod to include new liquidity tests
- Add TS compute-unit benchmark harness and generated output (tests/liquidity_math_cu_benchmarks.test.ts + out JSON)
- Minor fixes/tweaks to existing integration tests for consistency and precision
- Add documentation/results file for proptest regressions

This commit significantly expands testing, fuzzing, and benchmarking coverage for the liquidity math module.

+++ programs/fluxa_core/tests/core_arithmetic_integration_tests.rs
@@ -395,7 +395,7 @@ fn chained_high_precision_operations() {
     let mut expected = start;
 
     for &factor in &factors {
-        let scale = Q64x64::from_int(factor as u64);
+        let scale = Q64x64::from_int(factor);
         accum = accum.checked_mul(scale).unwrap();
         let rooted = sqrt_x64(accum).unwrap();
         let sqrt_scale = sqrt_x64(scale).unwrap();
